requirements.txt
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
google-api-core==2.24.1
google-api-python-client==2.162.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.68.0
httplib2==0.22.0
idna==3.10
iniconfig==2.0.0
oauthlib==3.2.2
packaging==24.2
pluggy==1.5.0
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyparsing==3.2.1
pytest==8.3.4
python-dateutil==2.9.0.post0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.17.0
SQLAlchemy==2.0.38
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
